backend/requirements.txt
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.1
channels==4.0.0
channels-redis==4.1.0
drf-spectacular==0.26.5
psycopg2-binary==2.9.7
redis==5.0.1
celery==5.3.4
python-dotenv==1.0.0
requests==2.31.0
transformers==4.35.2
torch==2.1.1
numpy==1.24.3
Pillow==10.1.0
gunicorn==21.2.0
whitenoise==6.6.0
